// Backend/src/common/enums/message-enum.ts
export enum SuccessMessage {
  USER_CREATED = 'User created successfully',
  LOGIN_SUCCESS = 'User logged in successfully',
  OTP_VERIFIED = 'OTP verification completed successfully',
  OTP_RESENT = 'OTP resent successfully',
  OTP_SENT = 'OTP sent successfully',
  PASSWORD_RESET = 'Password reset successfully',
  PASSWORD_UPDATED = 'Password updated successfully',
  LOGOUT_SUCCESS = 'Logged out successfully',
  UPGRADE_REQUEST_SENT = 'Upgrade request sent successfully',
  REAPPLY_REQUEST_SENT = ' Reapply request sent successfully',
}

export enum ErrorMessage {
  USER_ALREADY_EXISTS = 'User already exists',
  PASSWORD_MISMATCH = 'Password is not matching',
  USER_NOT_FOUND = 'User not found',
  INCORRECT_PASSWORD = 'Incorrect password',
  ACCOUNT_NOT_VERIFIED = 'Please verify your account first',
  ACCOUNT_BLOCKED = 'Your account is blocked',
  NO_OTP_FOUND = 'No OTP found',
  INCORRECT_OTP = 'Incorrect OTP',
  OTP_EXPIRED = 'OTP has expired',
  INVALID_PASSWORD_RESET = 'Invalid password reset request',
  NO_PERMISSION_ADMIN = 'No permission. Only admins allowed',
  REFRESH_TOKEN_MISSING = 'Refresh token is missing',
  INVALID_REFRESH_TOKEN = 'Invalid refresh token',
  TOKEN_MISSING = 'Token missing',
  INVALID_TOKEN = 'Invalid token',
  INTERNAL_SERVER_ERROR = 'Something went wrong. Please try again later.',
  WAIT_ONE_MINUTE = 'Please wait at least one minute before requesting a new OTP',
}